ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
google-generativeai==0.7.2
httpx==0.27.0
pymupdf==1.24.5
python-multipart==0.0.9
pydantic==2.7.4
requests==2.32.3
